requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
jinja2==3.1.6
pdfplumber==0.11.6
PyMuPDF==1.25.5
pytesseract==0.3.13
Pillow==11.2.1
